requirements.txt
adamp==0.3.0
numpy==2.1.3
omegaconf==2.3.0
Pillow==11.0.0
# pytorch_lightning==2.3.3
scipy==1.14.1
# torch==2.4.0+cu118
# torchvision==0.19.0+cu118
